Code Cleanup
Gave the reaction subscriber and message update subscriber their own classes, in order to clean things up a bit. Also added some sass (read: messages) to certain cases, rather than just having the bot wordlessly do things. Up next is looking into deleting the embedded message after the user picks the item, but I might do that in another branch, as this one needs merging into develop.
