+++ src/main/java/com/glacier/discordbot/commands/OrdinaryPlayer.java
@@ -24,7 +24,7 @@ public class OrdinaryPlayer implements Command {
 		IChannel textChannel = event.getChannel();
 		if(voiceChannel == null)
 		{
-			new JoinUser().runCommand(event, arguments);
+			UtilsAndConstants.sendMessage(textChannel, "Okay, but you won't hear me while I'm outside of a voice channel.");
 		}
         
 		//Turn the args back into a string separated by space
@@ -33,7 +33,7 @@ public class OrdinaryPlayer implements Command {
         GuildMusicManager musicManager = UtilsAndConstants.getGuildAudioPlayer(voiceChannel.getGuild());
         //get the queue manager for the channel the bot's in
 
-        System.out.println("Playing item " + searchStr + " because of " + event.getAuthor().getName() + " at " + UtilsAndConstants.getCurrentTimestamp());
+        System.out.println("Queueing item " + searchStr + " because of " + event.getAuthor().getName() + " at " + UtilsAndConstants.getCurrentTimestamp());
         
         playerManager.loadItemOrdered(musicManager, searchStr,
     		new OrdinaryResultHandler(textChannel,musicManager,searchStr));
